[MOD] moved entities in their own folder
