+++ src/main/java/com/chatop/api/entity/GenericEntity.java
@@ -1,4 +1,4 @@
-package com.chatop.api.model;
+package com.chatop.api.entity;
 
 import java.util.Date;
