+++ src/main/java/com/chatop/api/entity/RentalEntity.java
@@ -1,4 +1,4 @@
-package com.chatop.api.model;
+package com.chatop.api.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
