+++ src/main/java/com/chatop/api/entity/UserEntity.java
@@ -1,4 +1,4 @@
-package com.chatop.api.model;
+package com.chatop.api.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
