+++ src/main/java/com/chatop/api/model/CustomUserDetails.java
@@ -8,6 +8,8 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import com.chatop.api.entity.UserEntity;
+
 public class CustomUserDetails extends UserEntity implements UserDetails {
 
     private Collection<GrantedAuthority> authorities =  List.of(new SimpleGrantedAuthority("USER"));
