+++ src/main/java/com/chatop/api/repository/MessageRepository.java
@@ -3,7 +3,7 @@ package com.chatop.api.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.chatop.api.model.MessageEntity;
+import com.chatop.api.entity.MessageEntity;
 
 @Repository
 public interface MessageRepository extends JpaRepository<MessageEntity, Integer>{
