+++ src/main/java/com/chatop/api/repository/RentalRepository.java
@@ -3,7 +3,7 @@ package com.chatop.api.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.chatop.api.model.RentalEntity;
+import com.chatop.api.entity.RentalEntity;
 
 @Repository
 public interface RentalRepository extends JpaRepository<RentalEntity, Integer>{
