+++ src/main/java/com/chatop/api/service/MessageServiceImpl.java
@@ -6,11 +6,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import com.chatop.api.entity.MessageEntity;
+import com.chatop.api.entity.RentalEntity;
+import com.chatop.api.entity.UserEntity;
 import com.chatop.api.exception.ResourceNotFoundException;
 import com.chatop.api.model.Message;
-import com.chatop.api.model.MessageEntity;
-import com.chatop.api.model.RentalEntity;
-import com.chatop.api.model.UserEntity;
 import com.chatop.api.repository.MessageRepository;
 import com.chatop.api.repository.RentalRepository;
 import com.chatop.api.repository.UserRepository;
